Pytorch/requirements.txt
seaborn
matplotlib
pillow
torch
torchvision
tqdm
scikit-learn
pyqt5
